requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
Django==4.2.7
django-crispy-forms==2.1
djangorestframework==3.14.0
fontawesomefree==6.4.2
idna==3.6
Pillow==10.1.0
pytz==2023.4
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
